requirements.txt
Django==2.2.1
django-rest-framework==0.1.0
djangorestframework==3.9.4
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.2
pytz==2019.1
sqlparse==0.3.0
